Remember user to import Icon Library
When i read the readme it was not clear to me that i had to download or search for icon libraries to add to the head of the page, i though this was already built in with the project (like an import link or something). 
I Just mentioned Font awesome and Material design because those are the ones found when looking at configurations
```cs
<BlazoredToasts Position="ToastPosition.BottomRight"
                        Timeout="10"
                        IconType="IconType.[HERE]"
                        SuccessClass="success-toast-override"
                        SuccessIcon="fa fa-thumbs-up"
                        ErrorIcon="fa fa-bug" />
```

Maybe this is obvious for others but here is my suggestion to avoid the same problem i had. 

Thanks! :smile: